move send digits event types to a dedicated file
